Fix: minor changes
